Fixed most of the bugs with data persistence

+++ app/src/main/java/md/webmasterstudio/domenator/data/db/entity/ReportInfoEntity.kt
@@ -7,7 +7,7 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "reports_info")
 data class ReportInfoEntity(
     @PrimaryKey(autoGenerate = true) val id: Long = 0,
-    @ColumnInfo(name = "car_id") val carId: Int?,
+    @ColumnInfo(name = "car_id") val carId: Long?,
     @ColumnInfo(name = "date") val date: String?,
     @ColumnInfo(name = "fuel_amount") val fuelAmount: Float?,
     @ColumnInfo(name = "speedometer_value") val speedometerValue: Long?,
